move generation logic to it's own file
